git-svn-id: http://svn.code.sf.net/p/distributedcl/code/trunk@195 cb344d40-e590-48b6-843f-7a98a4fc1ff3

+++ src/libdistributedcl/remote/remote_platform.cpp
@@ -44,7 +44,7 @@ generic_context* remote_platform::create_context( const devices_t& devices ) con
 {
     dcl_message< msgCreateContext >* msg_ptr = new dcl_message< msgCreateContext >();
 
-    msg_ptr->set_device_count( devices.size() );
+    msg_ptr->set_device_count( static_cast<uint32_t>( devices.size() ) );
 
     for( devices_t::const_iterator it = devices.begin(); it != devices.end(); it++ )
     {
